Don't add line file info to ftable calls
Function calls via the ftable were having line/file arguments added to
them by insert line numbers, but we wouldn't actually pass the extra
args to the PRIM_FTABLE_CALL. This would cause us to use whatever
garbage was in our registers as line numbers and filenames, which
causes all sorts of problems when we are trying to use these filenames
in a lookup table. After this patch we now get new line and file info
inside of functions called via the ftable rather than have it passed in.