Allow some test systems file to be executable
Ordinarily test files such as .compopts, .execopts, or .execenv files are just
read in, which doesn't allow them to differ across configurations. Most of our
test systems files are read in with ReadFileWithComments() function. This adds
code to that function to read the output of a file if its executable instead of
just reading in the file.

This patch is motivated by wanting to turn off guard pages with an execenv, but
only for KNC.

Looking through sub_test I think this allows the following files to be
executable:

COMPOPTS    and .compopts
EXECOPTS    and .execopts
TIMEEXEC    and .timeexec
EXECENV     and .execenv
CHPLDOCOPTS and .chpldocopts
TIMEOUT     and .timeout
NUMLOCALES  and .numlocales
NUMTRIALS   and .numtrials

But it does not allow the following files to be executable yet:

.good        or .bad          (`diff` is called on the actual files)
LASTCOMPOPTS or .lastcompopts (`cat` is used to read in for some reason)
LASTEXECOPTS or .lastexecopts (`cat` is used to read in for some reason)
STDIN        or .stdin        (simply piped in with '<')
CATFILES     or .catfiles     (`cat` is used here)