Fix the moveDir use of copyTree slightly, and add a couple more tests
moveDir wasn't setting copySymbolically to true when it called copyTree, so
any symbolic links would be dropped on the floor.  Fixed that, and added tests
of a larger directory tree, and one that contains symbolic links in it.  Ensured
that all tests work with the copyTree symbolic link fix merged in from master.