Change includes so re2-interface.cc can compile with clang
The recent change to use C99 complex types causes re2-interface.cc to
choke when built with clang. There is an incompatibility when using
C99's complex.h and compiling C++ code with clang. Rather than try to
really figure out the root cause, or do a clean fix, I've just #ifdef'd
out the definitions in chplcast.h and chpltypes.h when compiling with
C++. A more principled approach might be to make it so re2-interface.cc
never includes these files in the first place, but that seemed to
require editing far to many include files.

Something to note is that a "convention" used in re2-interface.cc is not
actually valid code: placing imports inside of a `extern "C" {` block is
not allowed.