Fix order of initialization issue from replicateGlobalRecordWrappedVars
This PR modifies the order of initialization of arrays. It used to be
that something like
  var A: [1..3] real = (1.1, 2.2, 3.3);

would initialize A like this:

 addr = addr of type_tmp
 convertValueToRuntimeType(..., addr)
 op= (type_tmp, ...)
 PRIM_MOVE A, type_tmp

but after this PR's changes, the initialization sequence changed
slightly:

 addr = addr of A
 convertValueToRuntimeType(..., addr)
 op= (A, ...)

Before and after this commit, in the cases I saw,
replicateGlobalRecordWrappedVars adds the replication call
after a *block statement* doing the initialization. This commit
just disables some asserts that seem not to be relevant.