Add some error checking to the code that executes test system files
#2898 caused a bunch of nightly testing to fail because some test system files
have execute permissions by mistake. That said, my code wasn't very resilient
to these sorts of issues, and caused sub_test to except which ultimately led
start_test to exit because it didn't know how to handle the error which caused
jenkins to report a build failure.

In attempt to fail more gracefully and report a more helpful error message,
this adds some code to catch OSError's and report them.