Fix array initialization overflow bug
Previously, our array init code did something like:

    for i in 1..size {
      arr[i-1] = <init>
    }

However, if you had an array like:

    var foo: [0:uint(8)..254:uint(8)] int;

you'd end up with:

    for i in 1:uint(8)..255:uint(8) {...}

which isn't valid according to our current range overflow semantics. This patch
changes the init code to do something like:

    if size > 0 {
      for i in 0..size-1 {
        arr[i] = <init>
      }
    }

which is arguably better anyways since you avoid the `-1` each iteration. Note
that ideally I would have used 0..#size instead of checking that size > 0 and
then doing 0..size-1, but we don't optimize unbounded ranges that use the count
operator very well.

Note that a "maximal" arrays like:

    var foo: [0:uint(8)..255:uint(8)] int;

are still broken. This is for a slightly different reason. When we init arrays
we pass in `range.length` as the size of the array. Unfortunately range.length
overflows for maximal uint ranges, and for maximal/2 int ranges. For uint
arrays we end up passing in 0 as the length, which means the array is not
getting initialized.

I'm not quite sure what to do about that yet. The reason for the overflow is
because the return type of range.length is the type of the range, and no matter
how hard you squeeze the bits in 256 (length 0:uint(8)..255:uint(8)) just
doesn't fit in a uint(8)