Support building re2 with the cray compiler
Historically we have not supported building re2 with the cray compiler. This
was primarily because re2 throws gcc specific flags that the cray compiler
doesn't support. It's also possible that before the cray compiler threw -hgnu
by default it might not have been to build re2 even if it wasn't given bogus
flags.

This patch just overrides re2's default LDFLAGS and RE2_CXXFLAGS to not include
flags that are not supported or not needed for the cray compiler.

This is not a particularly elegant solution, it's mostly a proof of concept and
hopefully somebody more familiar with re2 can help me clean it up.

While working on this, I was wondering why we don't use our CXXFLAGS like we do
for our other third party libraries. It seems to me that we're building without
optimizations turned on. I think we're overriding re2s CXXFLAGS (which threw
-O3) but never asking for optimizations ourselves.