Remove the chplenv variable CHPL_THREADS
CHPL_THREADS used to be a full fledged chplenv variable that had user settable
options. However, 6a49e8876bd02f8dab6b18c97337a6e90a8b29e4 removed the ability
to set CHPL_THREADS independently of CHPL_TASKS.

That commit also added a warning to the script that inferred the value of
CHPL_THREADS to let users know that the variable is no longer user settable.
It's been long enough now that we don't believe such a warning is necessary and
as such there's no reason for CHPL_THREADS to remain as a chplenv. The only
thing it was used for anymore was to get the CHPL_MAKE_THREADS variable. This
commit removes all the code associated with CHPL_THREADS and defines
CHPL_MAKE_THREADS for each tasking layer in make/tasks/Makefile.<CHPL_TASKS>

This was motived by the recent chplenv rework that required special workarounds
for CHPL_THREADS.