In LLVM code generation, do not make inbounds GEPs
Our code generator was creating GetElementPointer instructions
with the "inbounds" attribute. This attribute adds error checking
that is not present for the C backend. Therefore, I have removed
it in order to improve LLVM backend performance.

It's as simple as changing the function we are calling
when creating these instructions (for e.g. array access).
 CreateInBoundsGEP -> CreateGEP