Include C compiler flags when linking for --llvm
When using --llvm, we link binaries with the clang++ command (as opposed
to a system linker like ld) so that clang++ can run link-time
optimizations. That is required with the current flow through the
compiler since we generate LLVM IR and then pass that to clang++ for code
generation. (A typical system linker would not know what to do with LLVM
IR).

Before this change, we were not actually running optimization or setting
the target architecture for the "link" call to clang++. That meant that
the LLVM backend was losing important optimization opportunities - in
particular, it was not vectorizing in certain simple cases that I
experimented with.

This change adds the optimization flags to the clang++ call by adding all
of the C compilation flags. This change gives approximate performance
parity for the C backend and the LLVM backend on a few simple loop/array
tests.