Fix llvm perf testing
We wanted this to be able to compare against the regular chapcs numbers, so it
needs to pretend to be chapcs as well. That way all the .dats wil be stored in
a subdirectory of chapcs which the perf compare scripts will then find.

Currently the llvm .dat files are going into:

    /data/sea/chapel/NightlyPerformance/chapcs-playground/llvm/

but we really just want them in:

    /data/sea/chapel/NightPerformance/chapcs/llvm/