Support building re2 with the cray and pgi compilers
Historically we have not supported building re2 with the cray or pgi compilers.
This was primarily because re2 throws gcc specific flags that these compilers
don't support.

This patch just overrides re2's default LDFLAGS and RE2_CXXFLAGS to not include
flags that are not supported or not needed for cray or pgi.