Removing a race from Tim's test_glob test
Tim was simultaneously iterating over an associative domain and
modifying it which is a race (and one that we can't detect).
We've somehow avoided this becoming a problem, though in Kyle's
testing it revealed itself.

Here, I fix this by making a copy of the associative domain
before iterating over it, which should be equivalent since
the whole thing is in an infinite loop until the domain is empty.

While, here, I changed do...while(true)'s into while(true)'s
which seems more readable to me.