Disable libudev in hwloc builds
Similarly to libnuma we're trying to statically link in libudev even though the
system only has a .so which is causing issues. Also like with libnuma the real
solution here is probably to only statically link in our third party libs and
let any dependencies be linked in dynamically.

libudev was added as a dependency with a recent version of hwloc which is why
it started causing issues when upgrading hwloc.