Fix bug with isHomogenousTupleType
A user discovered a bug where isHomogeneousTupleType assumed that it would only
be passed tuple types and would generate an internal error if passed a type
that wasn't a tuple.  This seemed like an unnecessary limitation, so I modified
the compiler definition of the primitive PRIM_IS_STAR_TUPLE_TYPE to stop
assuming this and instead return false when passed something that wasn't a
tuple.  While I was there, I also modified isHomogenousTupleValue to accept
an argument that wasn't a tuple, since otherwise we would get an error that the
function wasn't defined.

I could have redefined the module definition of isHomogeneousTupleType to make
the check before calling the primitive PRIM_IS_STAR_TUPLE_TYPE, but that would
not have fixed isHomogeneousTupleValue and would have required code duplication
to do so.