Undo/fix a couple of changes
I had mistakenly given the argument to the UseExpr constructor the same name as
the field I wished to assign to, which caused some initialization errors.  I
had also created an unused variable in my modification of ipeUtils when I really
just wanted to check on the type of the Expr provided.  Furthermore, discussion
with Mike indicated that the best strategy for dealing with
PRIM_USED_MODULES_LIST and the modUses field of BlockStmts was to merely adjust
them for using UseExprs instead of altering their type/definition more
drastically.  This meant that a couple of my changes needed to be backed out,
such as the modification of the modUses field's type and BlockStmt's
replaceChild method.