Get hwloc building with CCE
hwloc doesn't build with cce 8.4 or newer (and older versions if you throw
-hgnu, which is the default for 8.4)

It fails with:

    Declaration is incompatible with "int ffsl(long)" (declared at line 526 of
            "/usr/include/string.h").

There's a thread on the hwloc mailing lists about this issue that never really
got resolved. The thread starts with:
https://www.open-mpi.org/community/lists/hwloc-users/2013/01/0813.php

The main summary is that they believed it's a CCE bug, but it doesn't seem like
a bug was ever filed. I'll work on filing a bug with CCE, but want to get hwloc
building the in the meantime.

To work around this bug, this patch defines HWLOC_HAVE_BROKEN_FFS, which avoids
running into the problem code.

There's another possible issue where the vendor compiler is being detected as
gnu instead of cray. This is because hwloc checks for __GNUC__ before _CRAYC
and as of 8.4 (or earlier with -hgnu) cce defines __GNUC__ to provide gnu
extensions. A (hacky) workaround for this is to add

    CHPL_HWLOC_CFG_OPTIONS += hwloc_cv_c_compiler_vendor=cray

to our makefile, which will trick the configure script into thinking the
compiler vendor was cached. This is incredibly fragile/lame, but it seems to
work. When I was trying to get hwloc to build a few months ago I thought some
gnu specific flags were being thrown and causing other compilation error, but
that doesn't seem to be happening now. Still, I'd like to file a bug or submit
a patch to hwloc to fix this detection. It's a pretty easy fix, the check for
cray just has to be moved earlier (similar to what they already do for intel.)