Default to qthreads+hwloc with CCE
CCE 8.4 added support for gnu style inline assembly, which means we can now
build qthreads with it. This patch updates chplenv to default to qthreads if
we're using a new enough CCE (or if the module is being used since we know the
module is built with a new enough version of CCE.)

#2496 was an attempt to default to qthreads w/o hwloc for cce since hwloc
didn't build. However, hwloc is needed to get the best performance out of
qthreads. 792d4b37663ae5016075dab8f1b3ab0b5529a60c allows us to build hwloc
with CCE so we can now enable qthreads+hwloc and get better performance. For
stream-ep on a CRAY-XC this takes us from 65 GB/s with fifo to 74 GB/s.