Update BlockStmt's modUses field interactions to use new UseExpr AST node
This will preserve the except/only information given in a use statement when
the AST is transformed during scope resolution.  The method on a BlockStmt to
add used modules to its modUses field, moduleUseAdd, has been updated to
add UseExprs instead of ModuleSymbols.  Scope resolution's processImportExprs
was updated to traverse all UseExprs instead of all CallExprs.  There was one
hiccup from this effort, which was that because modUses now stores UseExprs,
processImportExprs encounters a use of the root module where previously this
use would have been considered implicit to the module itself and was inserted
directly into the block statement in question - to fix that, processImportExprs
was updated to ignore any encounters of the rootModule (though there should be
only one.  Any of them would encounter this error, and anyways, we don't
need to resolve the root module).