Whoops, missed a CallExpr to UseExpr case
That was because it was more related to PRIM_USED_MODULES_LIST.  Now instead
of the CallExpr for that primitive being the direct parent of SymExpr-wrapped
modules, there's a UseExpr in between.  Update that check for our new
expectations (hopefully that's all that's needed, though I'm not certain it is).