Modified three of tests w.r.t. our nightly testing.
* test/distributions/dm/hplx.chpl

While I had been adapting .bad to the test's various failure
behaviors, it continued being creating coming up with new ways to fail
and new output to print in those cases.
Following a request at today's group meeting, I am removing .bad,
hopefully eliminating the noise in testing.

* test/parallel/forall/vass/ri-3-stress.chpl

Now that we started XC testing on chpbld02, this test started
timing out there, just like it used to on chpbld01.

I am extending .skipif to skip it on chpbld02, in addition to chpbld01.
Suggested by Elliot.

* test/studies/hpcc/HPL/vass/hpl.hpcc2012.chpl

I am enabling this test to be run in our correctness configurations.
Before it was run only as a performance test on XC. Recently it had
an issue, Ben A went to investigate and had difficulty. He suggested
to setting it up for regular correctness testing to mitigate that difficulty.

[These tests are unrelated, I combined them to have just one PR, not three.]