Fixed a bug in reduce intent handling.
This PR fixes the following issue. Consider the parallel iterator
that is used in a forall loop with reduce intent(s). Consider the case
that the iterator has a yield inside an on-statement, and that
on-statement is not combined (e.g. via the "begin+on" optimization)
with a task-parallel construct.

My implementation of forall intents used to handle such on-statement
as if it were a task, and tried to apply task intents on it.
However, that is not right: on-statements by themselves are not subject
to task intents.

This PR modifies the way my implementation recurses into task-function-like
FnSymbols. Those not subject to task intents are handled using
the newly-introduced recursive propagateRecursively().
The other ones are handled using recursive propagateExtraLeaderArgs()
- no change there.

I am adding two tests that exercise this change:

  ri-iterator-with-on.chpl is perhaps the smallest setup, which I got
  by trimming down the other test.

  binary-tree-spawn.no-recurse.chpl is what I wrote to simulate
  binary tree-based spawning as a replacement for
  "for loc in Locales do on loc" without using a recursive iterator.
  (It works properly only for 1, 3, 7, or 15 locales.)

Separately - it seems like we do not perform the "begin+on" optimization
for the case where on-statement is a task in a cobegin.
To be addressed separately.