Parser set up for except keyword
Adds a parser rule for "use Foo except bar;", and constructs a UseExpr which
is aware of the excluded identifier.  I needed to change the UseExpr
constructor when a Vec is passed to not delete its contents.  This
implementation is not necessarily ideal, but it is a good starting point.  I
might switch to the std::vector constructor instead, or modify the parsing
rule to expect a more limited set of options for the tokens following the
except keyword.

This implementation does not yet respond to the excluded symbols, but it does
enable two of my tests to pass: the syntax error test is now correct, and the
verification that excluding a symbol does not bar other symbols from that
module passes (though this portion of the implementation is not where I would
expect a failure to arise).