Modify use of the new ChapelStringLiterals module
When I merged the string changes into my branch for adding the except keyword,
this line in symbol.cpp was causing me problems.  I determined that having
the compiler insert this use was unnecessary, so I removed its addition there
and added a use statement to ChapelBase (which was where it was getting
inserted).  This will avoid unnecessary fragility in my current branch to
otherwise fix the problem.