Add support in codegenAdd/Sub for ptr + int for LLVM
The new strings expect code generation to do something
intelligent with e.g.
 ptr + index
as a PRIM_ADD/PRIM_SUBTRACT. (where ptr is int8* index is int, e.g.).

However, this pattern was not supported with the LLVM backend.  This
patch adjusts codegenAdd and codegenSub to create the appropriate
getelementptr instruction in this case.