Updating tuple-of-string test for new strings
This test was originally written to improve and lock in an error
message indicating that iteration over strings was not permitted.
Now that it is permitted, the test "works", but generates an error
because the strings are of different lengths.  Yet, when running
--fast, such errors are disabled, so this test failed on fast
testing last night.

To make the test stable, I made the strings the same length
(sorry to my ex-colleagues for abbreviating your names to match
the four characters of 'brad').

Meanwhile, I introduced a new test (tupleOfNonIterableTypes.chpl)
to lock in the behavior that the tupleOfStrings test was originally
designed to test -- iterating over a tuple of types that did not
support iteration.