Remove fasta-lines suppressif
We used to suppress the fasta-lines performance test on machines that had < 8GB
of memory because of how much we leaked strings. Now that strings are records
(#2884) this test is passing on chap03, bradc-lnx, and the shootout box so we
can retire the suppressif! I also update the JIRA issue.

As a related note the performance of fasta-lines dramatically improved. It went
from 21.5 seconds down to 8.3 seconds on chap04