Alternate addition of use for the ChapelStringLiterals module
When I merged the string changes into my branch for adding the except keyword,
this line in symbol.cpp was causing me problems.  Kyle's preference was to
continue to insert the use of this module in the compiler instead of creating
a reference to a file that the reader cannot find in module space.  This change
will avoid unnecessary fragility in my current branch to otherwise fix the
problem.

Modified the output of test/modules/sungeun/init/printModuleInitOrder,
as the additional use statement changed its output.

Tested over std/