Fix completion processing for REPLY_PUT operations.
This fixes two problems and slightly improves the code.

The first problem was that for a REPLY_PUT, the info block "ack" member
wasn't initialized with the address of the done counter as it needed to
be.  This now fixed.

The second problem was that in AM_reply_put() we were responding with a
completion signal that was piggybacked on a gasnet_AMReplyLong2().  That
signal was delivered to the usual AM_signal() handler.  But that handler
is only appropriate for gasnet_AMReplyShortM() calls.  Handlers for
gasnet_AMReplyMediumM() (we don't use this) and gasnet_AMReplyLongM()
have different prototypes.  Thus the arguments for the handler were
incorrect, and the done counter manipulations caused segfaults.  The fix
is to add a new SIGNAL_LONG completion handler with the right prototype
for gasnet_AMReplyLongM() calls, and use that instead.

Finally, I also replaced the preprocessor #defines for the AM handler
function indices with an enumerated type.  I did this because I got a
bit hasty when adding the new SIGNAL_LONG handler index and left it with
the same value as the SIGNAL one, which GASNet (and bless it for doing
this) barked at when the code tried to install it.  Using an enumerated
type gives us the same values but with automatic incrementing, which
would prevent this sort of error.

The need for this resulted from yesterday's string-as-record change.
Apparently we have never before encountered an out-of-segment remote
address on a GET with GASNET_SEGMENT=fast before.