Change the locality of string relational operators
The previous string relational operators expected their right and left
hand sides to be on the same locale, and would go to that locale to
perform the comparison.  But, an autoCopy would localize lhs and rhs,
before they are placed into the on bundle, causing us to access garbage
data inside the locale portion of the comparisons routine.

This patch makes it so we will always localize the data, and perform the
comparison on the starting locale.