Close out numa, no-local, and linux32 regressions introduced last night
This commit updates .good files and suppressions for numa, no-local, and
linux32 as needed by yesterday's string commit to quiet them.

The .suppressif changes are logical given the failures last night yet,
unfortunately, are problematic when testing on Macs because this test
does not fail there (Similarly, plusEquals passed on gasnet.darwin
last night and reported a passing suppression as a result).  Here, I'm
relying on the fact that we don't do nightly no-local/numa testing on
Mac to rationalize this change.  If Mike didn't have his short-term
sights set on this test, I'd be more reluctant to add these lines
(alternatively, if we had boolean and we could suppress for linux as
well, but survey says we don't; and doing an executable skipif seemed
like overkill given the previous caveats).