Fix errors in realPath arising from the string changes on Tuesday
Something about the way the C function realpath creates the buffer it returns
was reacting poorly with gasnet fast and the new string set up.  Since the
locale method cwd() did not experience the same error, I surmised that it was
better to follow cwd()'s buffer creation strategy and use qio_malloc to create
a buffer to send to realpath.  This resolved the errors I was seeing.

I'm not entirely certain why our string changes tickled this, others might have
a better idea.