Retire Sung's c_string intents future
With the strings changes, this test seems to be working as we'd like
to me, though it's different than the .good predicted.  Specifically,
the test was trying to pass non-lvalue strings into routines with
'in', 'inout', and 'ref' intent, which should not be legal.  What
I've done here is to add a config param to look at different compile-time
failure modes while letting the cases that work as expected through.

Removed the .bad and .future and added .good files for the various
error cases.