Retire modNamedNewStringBreaks.chpl
With the recent reorganization of modules, this test no longer fails
and we don't seem to have any other standard modules that would
replicate the same failing behavior.  The case was very specific to
having ChapelStandard.chpl use a file in modules/standard/ that
defined two modules, one of which matched its filename and also the
name of a module defined by the user test code.  Currently, the only
standard modules that meet this requirement are not used by
ChapelStandard.chpl, and generally, we're trying to reduce the number
of files named in ChapelStandard.chpl, not grow it, so it doesn't make
sense to add one just to keep this future alive given how difficult it
is to run into, so I'm retiring it instead.