Close pipes in qio_proc_communicate
Close stdout/stderr pipes after we read EOF from each,
respectively. Technically, I believe it is possible for
a program to print something to the other pipe once one
of these closes. Also, closing the pipe might clean
up resources sooner.

Although we close the file (the pipe), we leave the
channel open, since it has buffered up the data. We just
mark its maximum position so that it will not attempt
to read from the closed file.