Renamed numLocales test to avoid conflict with NUMLOCALES
Late-breaking catch:  Because of case-insensitivity on Mac, the
test numLocales.chpl was writing its output binary as numLocales
which overwrote the NUMLOCALES file used for testing and broke
downstream tests.  This commit renames the test to numLocalesVar.chpl
and updates the references to it in the documentation to avoid this
problem.