Modify scopeResolve to maintain the included/excluded symbols in a use
During symbol lookup, when scopeResolve was creating the list of modules to
traverse next in a scope it was ignoring the other information in the UseExpr.
I modified the list to take in UseExprs instead, so that information would
still be present.

When creating the list of modules to traverse in order, looking for this symbol,
we were adding any module seen to a list so that we wouldn't search a particular
module more than once.  However, with use statements that specify symbols to
only search or exclude, merely seeing a module in a use statement is not
enough.  The symbols visible from one use may not be the same as those present
in a later use, and if the first use was more limited, we would miss finding
a symbol that we should otherwise find.  I think this is the right decision,
but others should feel free to re-evaluate this choice.  Note that a broad
application of non-transitive uses would avoid falling into this path.