Add two tests of the except keyword behavior
These verify the behavior of the previous commit, specifically that portion
involving whether a module will be added again to the search list the second
time it is encountered if the first encounter involved an except or only clause.
If we change our minds on that behavior, they will need to be updated.