Implement scopeResolve's check against excluded symbols, and a test update
Before looking for a match in the module's symbol table, first check if the
use statement defines the symbol you are looking for in the exclude list, and
if so, skip checking this module.  This causes all but a couple of the tests
I wrote to pass.  The two remaining tests I am now reconsidering if their
expected behavior is appropriate.  I also updated the expected output for one
test that was still looking for a pre-private improvement error state.