Fix cce link/sungeun/static_dynamic regression
The --dynamic versions of this test started failing with cce when qthreads was
made the default. To fix this, I'm adding -lrt as a link arg to qthreads when
cce is the target compiler.

Looking at `cc -craype-verbose -static` and `cc -craype-verbose -dynamic` we
can see that the `-static` versions automatically adds a `-lrt`, but the
dynamic version does not. `-static` is the default so I think during qthreads
build time it detects that it does not need to link librt itself. Then when a
chapel executable is being dynamically linked librt is missing.

I believe this is a similar issue that we ran into with #1466 and the follow-up
#1470. I'm not sure why this isn't an issue for gnu, intel, or pgi.