Remove maximum file size for using MMAP to read a file
From the very early days of QIO, we had

     qio_too_small_for_default_mmap
     qio_too_large_for_default_mmap

that control whether mmap is used as the mechanism to read a file in
choose_io_method.

We don't mmap very small files because it would not be memory efficient
to do so (since memory mapping will use file size rounded up to page size
memory).

When qio_too_large_for_default_mmap was introduced, I believe it
controlled memory mapping the entire file in some situations. That has
since been given its own varibale to control it:

     qio_initial_mmap_max

If we are not mapping the entire file when opening a file, there is no
reason to avoid using mmap to achieve the I/O for large files since we
will memory map only 8M = qio_mmap_chunk_iobufs * qbytes_iobuf_size at a
time. Note that we will memory map a file with size <
qio_initial_mmap_max (8MB) when we open the file in qio_mmap_initial. The
reason we do that is so that the memory map can be shared by all channels
working with the file - it saves memory when many channels are used.

At the same time, the default value of qio_too_large_for_default_mmap,
which was meant to be 64 GB, causes different behavior on 32-bit systems
since the constant number overflows.

Therefore, this commit removes qio_too_large_for_default_mmap.