Fix QIO bug exposed by recordReader test
The bug was that in _qio_buffered_advance_cached, we were clearing out
the channel fast-path cache (really just a pointer to a memory region we
can write to) but we never set cached_start = NULL. Other code then
unconditionally set cached_cur if cached_start was set (for this
application) which led to incorrect reads.

The fix is simple: set cached_start = NULL in
_qio_buffered_advance_cached.  The QIO code for this fast-path cache
should always check that cached_start != NULL before using the fast path.