Add an error overload for type*param-int
A user at the SC15 tutorial was surprised when typing a tuple type
as 'real*2' didn't generate a useful result (either working or
generating a useful error message).

This commit adds an error overload for this case and shows that
users can still overload it if they want to.

At today's meeting, we discussed making the real*2 case an error by
default to keep a smaller number of ways of expressing homogeneous
tuples in the language / improve readability of programs, and agreed
that we would reconsider this decision if users revolted to it over
time.