Add our "traditional" change that allows invoking gdb on child processes.
This change allows us to invoke gdb in a separate xterm window for each
child of a Chapel program that uses the GASNet amudprun launcher.  There
was a "traditional" patch for this which we have carried over the last
several GASNet releases, but now the code has changed so that it can no
longer be applied.  Here I'm adding the equivalent functionality, but
using new code in a slightly different (later) spot in the logic flow.