Adjust expiringValueOptimization to handle early returns
Since FLAG_INSERT_AUTO_DESTROY applies once per variable,
the optimization was incorrectly removing that flag when
a function might have control flow that does not go through
the block where the optimization was applied.

This problem was revealed by the test

    test/types/records/noakes/return.chpl

which has several test cases showing early returns from
a function.

To resolve the issue, only apply the optimization to variables
that are only written to in the basic block in which the
optimization would otherwise apply.