Work around acquiring internal Qthreads info we can't rewrite.
The tasking shim needs two pieces of information related to which
scheduler is being used:
- Does this scheduler allow enabling remote caching?  This is equivalent
  to asking whether the scheduler can move tasks (qthreads) from one
  worker pthread to another, since the remote caching relies on TLS for
  its caches.
- Does this scheduler run just one worker per shepherd, or more than
  that?  We need this information to decide how to set the QT_* env vars
  to parameterize QThreads behavior.

We can't get the answers to either of these through the API.  So in this
change, have the third-party Makefile produce a Chapel-specific .h file
(qthread-chapel.h) in the unique install subdir, containing #defines for
the answers to these two questions.  Then, #include that into the shim
and use the preprocessor-defined values where we need those answers.

While here, in the Makefile I also turned off building the internal (to
libqthread) version of the tasking shim, and reformatted the configure
command line not to overflow its line.