Update scopeResolve implementation to verify identifiers in except list
If an identifier in an except list corresponds to a private or non-existent
symbol in that use chain, the user likely made a mistake when typing the list.
To aid in debugging, inform them when this occurs so that they can fix it
right away.  Also update the tests of this behavior for a more precise error
message.

Details: As each use statement is processed, traverse its list of excluded
identifiers and look them up.  If an identifier is not able to be resolved,
error out.  If the match returned is a private symbol that is not visible in
this context, also error out.